15-05-2024

+++ CustomArrayList/Main.java
@@ -4,7 +4,8 @@ package CustomArrayList;
 public class Main {
     public static void main(String[] args) {
 
-        WildCard<Float> list = new WildCard<>();
+        // WildCard<String> listStr = new WildCard<>(); //Now you can create anytype with it
+        WildCard<Integer> list = new WildCard<>();
         for (int i = 0; i < 14; i++) {
             list.add(i);
         }
